gl2: make cleanup of resources effective again.
Prior to this commit resources were only cleaned up at a call to Render
each frame. This worked fine in v1 as we used a single window. Now that
we are using multiple windows and keeping our resources in a fake
gl2.Device nobody ever calls Render on that device, and resources would
be entirely leaked without this change.